Algae_Antioxidant_Project/requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
seaborn>=0.13
matplotlib>=3.7
statsmodels>=0.14
xgboost>=2.0
shap>=0.44
scipy>=1.11
pyyaml>=6.0
torch>=2.2; platform_system!='Windows' or platform_machine!='ARM64'
torch>=2.2; platform_system=='Windows' and platform_machine!='ARM64'
